# api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "hivent-users-api"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "users"
paths:
  /users/{id}:
    get:
      tags:
      - "users"
      operationId: "readUser"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            $ref: "#/definitions/User"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Users"
    put:
      tags:
      - "users"
      operationId: "updateUser"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Mise à jour réussie."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Erreur lors de la mise à jour."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Users"
    delete:
      tags:
      - "users"
      operationId: "deleteUser"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Users"
  /users:
    get:
      tags:
      - "users"
      operationId: "findUsers"
      parameters:
      - name: "email"
        in: "query"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Lecture réussie."
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Users"
    post:
      tags:
      - "users"
      operationId: "createUser"
      parameters:
      - in: "body"
        name: "user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/User"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Users"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Identifiant.Lecture seule.\n"
      name:
        type: "string"
      email:
        type: "string"
      lastConnection:
        type: "string"
        format: "date-time"
    example:
      name: "name"
      id: 0
      email: "email"
      lastConnection: "2000-01-23T04:56:07.000+00:00"
  ReponseErreur:
    type: "object"
    properties:
      erreurs:
        type: "array"
        items:
          $ref: "#/definitions/ReponseErreur_erreurs"
  ReponseErreur_erreurs:
    properties:
      champErreur:
        type: "string"
      codeErreur:
        type: "string"
      libelleErreur:
        type: "string"